f32

+++ ass2/src/lenia_cpu.c
@@ -37,28 +37,28 @@ struct orbium_coo {
 };
 
 // Function to calculate Gaussian
-inline f64 gauss(const f64 x, const f64 mu, const f64 sigma) {
-    return exp(-0.5 * pow((x - mu) / sigma, 2));
+inline f32 gauss(const f32 x, const f32 mu, const f32 sigma) {
+    return expf(-0.5f * powf((x - mu) / sigma, 2));
 }
 
 // Function for growth criteria
-f64 growth_lenia(const f64 u) {
-    f64 mu = 0.15;
-    f64 sigma = 0.015;
+f32 growth_lenia(const f32 u) {
+    f32 mu = 0.15;
+    f32 sigma = 0.015;
     return -1 + 2 * gauss(u, mu, sigma);  // Baseline -1, peak +1
 }
 
 // Function to generate convolution kernel
-f64* generate_kernel(f64* const K, const usize size) {
+f32* generate_kernel(f32* const K, const usize size) {
     // Construct ring convolution filter
-    const f64 mu = 0.5;
-    const f64 sigma = 0.15;
+    const f32 mu = 0.5;
+    const f32 sigma = 0.15;
     const int r = size / 2;
-    f64 sum = 0;
+    f32 sum = 0;
     if (K != NULL) {
         for (int y = -r; y < r; y++) {
             for (int x = -r; x < r; x++) {
-                f64 distance = sqrt((1 + x) * (1 + x) + (1 + y) * (1 + y)) / r;
+                f32 distance = sqrt((1 + x) * (1 + x) + (1 + y) * (1 + y)) / r;
                 K[(y + r) * size + x + r] = gauss(distance, mu, sigma);
                 if (distance > 1) {
                     K[(y + r) * size + x + r] = 0;  // Cut at d=1
@@ -94,9 +94,9 @@ int main() {
 #endif
 
     // Allocate memory
-    f64* w = (f64*)calloc(KERNEL_SIZE * KERNEL_SIZE, sizeof(f64));
-    f64* world = (f64*)calloc(N * N, sizeof(f64));
-    f64* tmp = (f64*)calloc(N * N, sizeof(f64));
+    f32* w = (f32*)calloc(KERNEL_SIZE * KERNEL_SIZE, sizeof(f32));
+    f32* world = (f32*)calloc(N * N, sizeof(f32));
+    f32* tmp = (f32*)calloc(N * N, sizeof(f32));
 
     // Generate convolution kernel
     w = generate_kernel(w, KERNEL_SIZE);
@@ -106,7 +106,7 @@ int main() {
         world = place_orbium(world, N, N, orbiums[o].row, orbiums[o].col, orbiums[o].angle);
     }
 
-    f64 start = omp_get_wtime();
+    f32 start = omp_get_wtime();
 
     // Lenia Simulation
     for (unsigned int step = 0; step < NUM_STEPS; step++) {
@@ -115,7 +115,7 @@ int main() {
         OMP(parallel for)
         for (usize i = 0; i < N; i++) {
             for (usize j = 0; j < N; j++) {
-                f64 sum = 0;
+                f32 sum = 0;
                 for (int ki = KERNEL_SIZE - 1, kri = 0; ki >= 0; ki--, kri++) {
                     for (int kj = KERNEL_SIZE - 1, kcj = 0; kj >= 0; kj--, kcj++) {
                         sum += w(ki, kj) * input((i - KERNEL_SIZE / 2 + N + kri), (j - KERNEL_SIZE / 2 + N + kcj));
@@ -144,7 +144,7 @@ int main() {
 #ifdef GENERATE_GIF
     ge_close_gif(gif);
 #endif
-    f64 stop = omp_get_wtime();
+    f32 stop = omp_get_wtime();
     printf("Time(full): %f s\n", stop - start);
     free(w);
     free(tmp);
